🐛 Fix issue with CoinMarketCap Connector
In the `coinmarketcap_connector.py` file, this commit fixes a bug where the CoinMarketCap Connector was not returning accurate cryptocurrency data. The bug was caused by a typo in the API endpoint URL, which has now been corrected. With this fix, the connector will now successfully retrieve the correct information from the CoinMarketCap API.